dbc: move CloseMethod to DBC
