+++ dbc/src/opret/mod.rs
@@ -28,7 +28,9 @@ mod spk;
 use bc::Tx;
 use commit_verify::mpc::Commitment;
 use commit_verify::{CommitmentProtocol, EmbedCommitVerify, EmbedVerifyError};
+use strict_encoding::{StrictDeserialize, StrictSerialize};
 
+use crate::proof::Method;
 use crate::{Proof, LIB_NAME_BPCORE};
 
 /// Marker non-instantiable enum defining LNPBP-12 taproot OP_RETURN (`tapret`)
@@ -67,8 +69,12 @@ pub enum OpretError {
 )]
 pub struct OpretProof(());
 
+impl StrictSerialize for OpretProof {}
+impl StrictDeserialize for OpretProof {}
+
 impl Proof for OpretProof {
     type Error = EmbedVerifyError<OpretError>;
+    const METHOD: Method = Method::OpretFirst;
 
     fn verify(&self, msg: &Commitment, tx: &Tx) -> Result<(), EmbedVerifyError<OpretError>> {
         tx.verify(msg, self)
