+++ dbc/src/proof.rs
@@ -21,17 +21,63 @@
 
 use std::error::Error;
 use std::fmt::Debug;
+use std::str::FromStr;
 
 use bc::Tx;
 use commit_verify::{mpc, CommitEncode};
-use strict_encoding::{StrictDecode, StrictDumb, StrictEncode};
+use strict_encoding::{StrictDecode, StrictDeserialize, StrictDumb, StrictEncode, StrictSerialize};
+
+use crate::LIB_NAME_BPCORE;
+
+/// wrong deterministic bitcoin commitment closing method id '{0}'.
+#[derive(Clone, PartialEq, Eq, Debug, Display, Error, From)]
+#[display(doc_comments)]
+pub struct MethodParseError(pub String);
+
+/// Method of DBC construction.
+#[derive(Copy, Clone, Ord, PartialOrd, Eq, PartialEq, Hash, Debug, Display)]
+#[cfg_attr(
+    feature = "serde",
+    derive(Serialize, Deserialize),
+    serde(crate = "serde_crate", rename_all = "camelCase")
+)]
+#[derive(StrictType, StrictDumb, StrictEncode, StrictDecode)]
+#[strict_type(lib = LIB_NAME_BPCORE, tags = repr, into_u8, try_from_u8)]
+#[repr(u8)]
+pub enum Method {
+    /// OP_RETURN commitment present in the first OP_RETURN-containing
+    /// transaction output.
+    #[display("opret1st")]
+    #[strict_type(dumb)]
+    OpretFirst = 0x00,
+
+    /// Taproot-based OP_RETURN commitment present in the first Taproot
+    /// transaction output.
+    #[display("tapret1st")]
+    TapretFirst = 0x01,
+}
+
+impl FromStr for Method {
+    type Err = MethodParseError;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        Ok(match s.to_lowercase() {
+            s if s == Method::OpretFirst.to_string() => Method::OpretFirst,
+            s if s == Method::TapretFirst.to_string() => Method::TapretFirst,
+            _ => return Err(MethodParseError(s.to_owned())),
+        })
+    }
+}
 
 /// Deterministic bitcoin commitment proof types.
 pub trait Proof:
-    Clone + Eq + CommitEncode + Debug + StrictEncode + StrictDecode + StrictDumb
+    Clone + Eq + Debug + CommitEncode + StrictSerialize + StrictDeserialize + StrictDumb
 {
     /// Verification error.
     type Error: Error;
+
+    /// Returns DBC method used by the proof.
+    const METHOD: Method;
 
     /// Verifies DBC proof against the provided transaction.
     fn verify(&self, msg: &mpc::Commitment, tx: &Tx) -> Result<(), Self::Error>;
