+++ stl/BPCore@0.1.0.sta
@@ -1,5 +1,5 @@
 -----BEGIN STRICT TYPE LIB-----
-Id: urn:ubideco:stl:5Nyd3BhfDAEfc5th2httogefZGzMftfcJNh5Lo9guuTx
+Id: urn:ubideco:stl:JAeer3PsTTQRrGttsnPXJKdBTJqQwQ5ZwWHiABEDoCyd
 Name: BPCore
 Dependencies: 
   urn:ubideco:stl:ZtHaBzu9ojbDahaGKEXe5v9DfSDxLERbLkEB23R6Q6V,
@@ -50,39 +50,39 @@ ZWVUYXByZXRQcm9vZgYDBHR4aWQC9WwTYiP2OadKCZPcR0bJ+YqruINYXbXZFj8Y
 fsQoGgqjgkLzy9fR0KES2o3hYC9W1PhvDsTEdsXAaFlMSwRlVghtcGNQcm9vZgII
 bJMpP1Zo7NnfnUB1CNehMyMWREFWAosurAm/5d+NQsuFp4sSpTdM//sj9cu/mbZk
 SymVhn1xF4kiNVn8e4tXCGRiY1Byb29mAQ+2H5g/Gu2rjnvK5hyt61m+s5sC5IXz
-N5lwiJbZEwgMDkJsaW5kU2VhbFR4UHRyBgQGbWV0aG9kAdJSMHx9cZA19zB+6Cgl
-7BuBK6GH3Z2YSYtKjRb6btVDBHR4aWQBseU4ORQKOL7wbxrPvhxgpW/h4fR9eAgo
+N5lwiJbZEwgMDkJsaW5kU2VhbFR4UHRyBgQGbWV0aG9kAS63ECW5bmUW2nsUHaEd
+nW9ZOnia/Ulmt3/A+t3UnJLrBHR4aWQBseU4ORQKOL7wbxrPvhxgpW/h4fR9eAgo
 eb/R+tMQId0Edm91dAL1bBNiI/Y5p0oJk9xHRsn5iqu4g1hdtdkWPxh+xCgaCiHj
 PkPFqlzyKSdTozjBZ+07Y5xN2c69qY80aRe6yUN1CGJsaW5kaW5nAAAIDUJsaW5k
-U2VhbFR4aWQGBAZtZXRob2QB0lIwfH1xkDX3MH7oKCXsG4EroYfdnZhJi0qNFvpu
-1UMEdHhpZAL1bBNiI/Y5p0oJk9xHRsn5iqu4g1hdtdkWPxh+xCgaCqOCQvPL19HQ
+U2VhbFR4aWQGBAZtZXRob2QBLrcQJbluZRbaexQdoR2db1k6eJr9SWa3f8D63dSc
+kusEdHhpZAL1bBNiI/Y5p0oJk9xHRsn5iqu4g1hdtdkWPxh+xCgaCqOCQvPL19HQ
 oRLajeFgL1bU+G8OxMR2xcBoWUxLBGVWBHZvdXQC9WwTYiP2OadKCZPcR0bJ+Yqr
 uINYXbXZFj8YfsQoGgoh4z5Dxapc8iknU6M4wWftO2OcTdnOvamPNGkXuslDdQhi
-bGluZGluZwAACAtDbG9zZU1ldGhvZAMCCm9wcmV0Rmlyc3QAC3RhcHJldEZpcnN0
-AQxFeHBsaWNpdFNlYWwGAwZtZXRob2QB0lIwfH1xkDX3MH7oKCXsG4EroYfdnZhJ
-i0qNFvpu1UMEdHhpZAL1bBNiI/Y5p0oJk9xHRsn5iqu4g1hdtdkWPxh+xCgaCqOC
-QvPL19HQoRLajeFgL1bU+G8OxMR2xcBoWUxLBGVWBHZvdXQC9WwTYiP2OadKCZPc
-R0bJ+YqruINYXbXZFj8YfsQoGgoh4z5Dxapc8iknU6M4wWftO2OcTdnOvamPNGkX
-uslDdRFFeHBsaWNpdFNlYWxUeFB0cgYDBm1ldGhvZAHSUjB8fXGQNfcwfugoJewb
-gSuhh92dmEmLSo0W+m7VQwR0eGlkAbHlODkUCji+8G8az74cYKVv4eH0fXgIKHm/
-0frTECHdBHZvdXQC9WwTYiP2OadKCZPcR0bJ+YqruINYXbXZFj8YfsQoGgoh4z5D
-xapc8iknU6M4wWftO2OcTdnOvamPNGkXuslDdQpPcHJldFByb29mBQEAAAAKU2Vj
-cmV0U2VhbAUBAAcAAEAgABFUYXByZXROb2RlUGFydG5lcgQDAAhsZWZ0Tm9kZQAF
-AQL1bBNiI/Y5p0oJk9xHRsn5iqu4g1hdtdkWPxh+xCgaCjG7ruiEiXfFsAcxcjHN
-kvYwtiktzYl/3LrexbrE6X/ZAQlyaWdodExlYWYABQEC9WwTYiP2OadKCZPcR0bJ
-+YqruINYXbXZFj8YfsQoGgpfrNlt5T9OxRTZsY1ngNoKq+stJgv9Ny2jwHqkXTmf
-TQILcmlnaHRCcmFuY2gABQEBOD9iLnFT0sghkTzLdx2fPWTfdvIoVVkt+EZDlBZN
-bQUPVGFwcmV0UGF0aFByb29mBgILcGFydG5lck5vZGUABAIABG5vbmUAAAABBHNv
-bWUABQEBAXwHXQBASxaVmBsXw2Etxrjqrg1/cFlikhZ1WuUZkxEFbm9uY2UAAAEL
-VGFwcmV0UHJvb2YGAglwYXRoUHJvb2YBE8RTUmYnu0QljDtn9MzCfv785Ce3z14P
-/YGPL3572HwKaW50ZXJuYWxQawL1bBNiI/Y5p0oJk9xHRsn5iqu4g1hdtdkWPxh+
-xCgaCqmFhHdTnmj7sL/ux5ohxFdDwVT4+r0/mODTkT3ntzd4EVRhcHJldFJpZ2h0
-QnJhbmNoBgIMbGVmdE5vZGVIYXNoAvVsE2Ij9jmnSgmT3EdGyfmKq7iDWF212RY/
-GH7EKBoKMbuu6ISJd8WwBzFyMc2S9jC2KS3NiX/cut7FusTpf9kNcmlnaHROb2Rl
-SGFzaAL1bBNiI/Y5p0oJk9xHRsn5iqu4g1hdtdkWPxh+xCgaCjG7ruiEiXfFsAcx
-cjHNkvYwtiktzYl/3LrexbrE6X/ZBVR4UHRyBAIACXdpdG5lc3NUeAAAAAEEdHhp
-ZAAFAQL1bBNiI/Y5p0oJk9xHRsn5iqu4g1hdtdkWPxh+xCgaCqOCQvPL19HQoRLa
-jeFgL1bU+G8OxMR2xcBoWUxLBGVW
+bGluZGluZwAACAxFeHBsaWNpdFNlYWwGAwZtZXRob2QBLrcQJbluZRbaexQdoR2d
+b1k6eJr9SWa3f8D63dSckusEdHhpZAL1bBNiI/Y5p0oJk9xHRsn5iqu4g1hdtdkW
+Pxh+xCgaCqOCQvPL19HQoRLajeFgL1bU+G8OxMR2xcBoWUxLBGVWBHZvdXQC9WwT
+YiP2OadKCZPcR0bJ+YqruINYXbXZFj8YfsQoGgoh4z5Dxapc8iknU6M4wWftO2Oc
+TdnOvamPNGkXuslDdRFFeHBsaWNpdFNlYWxUeFB0cgYDBm1ldGhvZAEutxAluW5l
+Ftp7FB2hHZ1vWTp4mv1JZrd/wPrd1JyS6wR0eGlkAbHlODkUCji+8G8az74cYKVv
+4eH0fXgIKHm/0frTECHdBHZvdXQC9WwTYiP2OadKCZPcR0bJ+YqruINYXbXZFj8Y
+fsQoGgoh4z5Dxapc8iknU6M4wWftO2OcTdnOvamPNGkXuslDdQZNZXRob2QDAgpv
+cHJldEZpcnN0AAt0YXByZXRGaXJzdAEKT3ByZXRQcm9vZgUBAAAAClNlY3JldFNl
+YWwFAQAHAABAIAARVGFwcmV0Tm9kZVBhcnRuZXIEAwAIbGVmdE5vZGUABQEC9WwT
+YiP2OadKCZPcR0bJ+YqruINYXbXZFj8YfsQoGgoxu67ohIl3xbAHMXIxzZL2MLYp
+Lc2Jf9y63sW6xOl/2QEJcmlnaHRMZWFmAAUBAvVsE2Ij9jmnSgmT3EdGyfmKq7iD
+WF212RY/GH7EKBoKX6zZbeU/TsUU2bGNZ4DaCqvrLSYL/Tcto8B6pF05n00CC3Jp
+Z2h0QnJhbmNoAAUBATg/Yi5xU9LIIZE8y3cdnz1k33byKFVZLfhGQ5QWTW0FD1Rh
+cHJldFBhdGhQcm9vZgYCC3BhcnRuZXJOb2RlAAQCAARub25lAAAAAQRzb21lAAUB
+AQF8B10AQEsWlZgbF8NhLca46q4Nf3BZYpIWdVrlGZMRBW5vbmNlAAABC1RhcHJl
+dFByb29mBgIJcGF0aFByb29mARPEU1JmJ7tEJYw7Z/TMwn7+/OQnt89eD/2Bjy9+
+e9h8CmludGVybmFsUGsC9WwTYiP2OadKCZPcR0bJ+YqruINYXbXZFj8YfsQoGgqp
+hYR3U55o+7C/7seaIcRXQ8FU+Pq9P5jg05E957c3eBFUYXByZXRSaWdodEJyYW5j
+aAYCDGxlZnROb2RlSGFzaAL1bBNiI/Y5p0oJk9xHRsn5iqu4g1hdtdkWPxh+xCga
+CjG7ruiEiXfFsAcxcjHNkvYwtiktzYl/3LrexbrE6X/ZDXJpZ2h0Tm9kZUhhc2gC
+9WwTYiP2OadKCZPcR0bJ+YqruINYXbXZFj8YfsQoGgoxu67ohIl3xbAHMXIxzZL2
+MLYpLc2Jf9y63sW6xOl/2QVUeFB0cgQCAAl3aXRuZXNzVHgAAAABBHR4aWQABQEC
+9WwTYiP2OadKCZPcR0bJ+YqruINYXbXZFj8YfsQoGgqjgkLzy9fR0KES2o3hYC9W
+1PhvDsTEdsXAaFlMSwRlVg==
 
 -----END STRICT TYPE LIB-----
